refactor: decompose Studio god components and standardize API routes
Phase 4 — Component decomposition:
- StudioWorkspace: 614 → 80 lines, extracts 6 focused sub-components
  (StudioCardSelectors, StudioPreview, StudioToolbarPanels,
   StudioCardManagement, StudioProjectHistory, StudioVideoMode)
- CharacterCardManager: 826 → 3 files
  (CharacterCardCreateForm, CharacterCardItem, orchestrator ~120 lines)
- All sub-components consume split contexts directly (zero prop drilling)

Phase 5 — API route standardization:
- Migrated generate, studio/generate, image/analyze to createApiRoute factory
- Migrated generate-long-video/cancel, retry to createApiRoute factory
- Cleaned up GET routes (generate-video/status, generate-long-video/status)
  with unified GenerationError handling pattern
- Added longVideoCancel/Retry/Status to RATE_LIMIT_CONFIGS

tsc --noEmit: 0 errors
