+++ src/constants/config.ts
@@ -270,6 +270,9 @@ export const RATE_LIMIT_CONFIGS = {
   studioGenerate: { limit: 10, windowSeconds: 60 },
   generateVideo: { limit: 5, windowSeconds: 60 },
   generateLongVideo: { limit: 3, windowSeconds: 60 },
+  longVideoCancel: { limit: 10, windowSeconds: 60 },
+  longVideoRetry: { limit: 5, windowSeconds: 60 },
+  longVideoStatus: { limit: 30, windowSeconds: 60 },
   imageEdit: { limit: 10, windowSeconds: 60 },
   imageAnalyze: { limit: 10, windowSeconds: 60 },
   promptEnhance: { limit: 20, windowSeconds: 60 },
